feat: 게임 시작 아이템 리스트 조회 기능 수정
- 아이템 랭킹 추가

+++ src/main/java/com/hoshogi/onlyonepick/domain/game/controller/GameController.java
@@ -54,7 +54,7 @@ public class GameController {
 
     @GetMapping("/{game-id}/items")
     ApiResponse<List<GameItemResponse>> showGameItems(@PathVariable("game-id") Long gameId,
-                                                      @RequestParam Long count) {
+                                                      @RequestParam int count) {
         return ApiResponse.onSuccess(OK, gameService.showGameItems(gameId, count));
     }
 }
